AutoVideoMiner/requirements.txt
# Core orchestration and typing
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0

# AWS Bedrock integrations
langchain-aws>=0.1.0
boto3>=1.34.0

# Config and environment
PyYAML>=6.0.1
python-dotenv>=1.0.1

# GUI
streamlit>=1.36.0

# Video search/download
yt-dlp>=2024.8.6
